Revert "Downgrade etcd Chart to 5.3.2"
This reverts commit a5d5c391c7b253f5563f32cca4496828449b0552.

The Yake `1.86` release upgrade its `etcd` chart from `5.3.2` to `6.0.0`. This performs an upgrade of `etcd` itself and a renaming of the `PVC`s.

In the `ske-base` [PR](https://github.com/stackitcloud/ske-base/pull/1000) will be added a `postrenderers` function and resets the renaming.

The commit "Downgrade etcd Chart to 5.3.2" can be dropped with the next
version 1.87.
This revert commit can be dropped with the next version 1.87.